LayoutTests/js/regexp-ranges-and-escaped-hyphens-expected.txt

Tests for bug #21232, and related range issues described in bug.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regexp01.toString() is "1235"
PASS regexp01a.toString() is "123 5"
PASS regexp01b.toString() is "1-3 5"
PASS regexp01c.toString() is "1-3 5"
PASS regexp01d.toString() is "123 5"
PASS regexp01e.toString() is "123 5"
PASS regexp01f.toString() is "-3"
PASS regexp01g.toString() is "2-"
PASS regexp02.toString() is "1-35"
PASS regexp02a.toString() is "1-3 5"
PASS regexp02b.toString() is "1-3 5"
PASS regexp02c.toString() is "1-3 5"
PASS regexp02d.toString() is "1-3 5"
PASS regexp02e.toString() is "1-3 5"
PASS regexp03a.toString() is "-.0"
PASS regexp03b.toString() is "+,-"
PASS bug21232 is false
PASS successfullyParsed is true

TEST COMPLETE
